renamed umich detector
