--- apriltag_detector_umich/.github/workflows/ci.yaml
@@ -1,24 +1,0 @@
-#
-# continuous integration workflow
-#
-name: build repo
-
-on:
-  push:
-    branches: [master]
-  pull_request:
-    branches: [master]
-  workflow_dispatch:
-    branches: [master]
-
-jobs:
-  build_ros1:
-    uses: ros-misc-utilities/ros_build_scripts/.github/workflows/ros1_ci.yml@master
-    with:
-      repo: ${{ github.event.repository.name }}
-      vcs_url: https://raw.githubusercontent.com/${{ github.repository }}/master/${{ github.event.repository.name }}.repos
-  build_ros2:
-    uses: ros-misc-utilities/ros_build_scripts/.github/workflows/ros2_recent_ci.yml@master
-    with:
-      repo: ${{ github.event.repository.name }}
-      # vcs_url: https://raw.githubusercontent.com/${{ github.repository }}/master/${{ github.event.repository.name }}.repos
