+++ apriltag_detector_umich/package.xml
@@ -1,29 +1,21 @@
 <?xml version="1.0"?>
 <package format="3">
-  <name>apriltag_detector</name>
+  <name>apriltag_detector_umich</name>
   <version>1.0.0</version>
-  <description>ROS package for apriltag detection</description>
+  <description>ROS2 package for apriltag detection with the UMich detector</description>
   <maintainer email="bernd.pfrommer@gmail.com">Bernd Pfrommer</maintainer>
   <license>Apache2</license>
 
-  <!-- ROS2 dependencies -->
-  <buildtool_depend condition="$ROS_VERSION == 2">ament_cmake</buildtool_depend>
-  <buildtool_depend condition="$ROS_VERSION == 2">ament_cmake_ros</buildtool_depend>
-  <buildtool_depend condition="$ROS_VERSION == 2">ament_cmake_auto</buildtool_depend>
-  <depend condition="$ROS_VERSION == 2">rclcpp</depend>
-  <depend condition="$ROS_VERSION == 2">rclcpp_components</depend>
-  <depend condition="$ROS_VERSION == 2">apriltag_msgs</depend>
-  <test_depend condition="$ROS_VERSION == 2">ament_lint_auto</test_depend>
-  <test_depend condition="$ROS_VERSION == 2">ament_lint_common</test_depend>
-  <test_depend condition="$ROS_VERSION == 2">ament_cmake_clang_format</test_depend>
+  <buildtool_depend>ament_cmake</buildtool_depend>
+  <buildtool_depend>ament_cmake_ros</buildtool_depend>
+  <buildtool_depend>ament_cmake_auto</buildtool_depend>
+  <depend>rclcpp</depend>
+  <depend>rclcpp_components</depend>
+  <depend>apriltag_msgs</depend>
+  <test_depend>ament_lint_auto</test_depend>
+  <test_depend>ament_lint_common</test_depend>
+  <test_depend>ament_cmake_clang_format</test_depend>
 
-  <!-- ROS1 dependencies -->
-  <buildtool_depend condition="$ROS_VERSION == 1">catkin</buildtool_depend>
-  <depend condition="$ROS_VERSION == 1">roscpp</depend>
-  <depend condition="$ROS_VERSION == 1">nodelet</depend>
-  <depend condition="$ROS_VERSION == 1">apriltag_detector_msgs</depend>
-
-  <!-- common ROS1/ROS2 dependencies -->
   <buildtool_depend>ros_environment</buildtool_depend> <!-- ROS_VERSION + ROS_DISTRO -->
   <depend>image_transport</depend>
   <depend>sensor_msgs</depend>
@@ -31,10 +23,7 @@
   <depend>apriltag</depend> <!-- the umich apriltag library -->
   
   <export>
-    <nodelet plugin="${prefix}/nodelet_plugins.xml"/>
-    <!-- this is crucial else the package will not be registered! -->
-    <build_type condition="$ROS_VERSION == 1">catkin</build_type>
-    <build_type condition="$ROS_VERSION == 2">ament_cmake</build_type>
+    <build_type>ament_cmake</build_type>
   </export>
 
 </package>
